moved GetSuppressionBBID method away from nfs code to api
